lrs-device-controller holds device streams
