+++ tools/dds/dds-server/lrs-device-controller.h
@@ -3,31 +3,32 @@
 
 #pragma once
 #include <librealsense2/rs.hpp>  // Include RealSense Cross Platform API
+#include <librealsense2/utilities/json.h>
+
 #include <unordered_map>
 
-namespace tools {
+namespace realdds {
+    class dds_device_server;
+}
 
+namespace tools {
 
 // This class is in charge of handling a RS device: streaming, control..
 class lrs_device_controller
 {
 public:
-
-    typedef std::function< void(
-        rs2::frame ) >
-        frame_callback_type;
-
-    lrs_device_controller( rs2::device dev );
+    lrs_device_controller( rs2::device dev, std::shared_ptr< realdds::dds_device_server > dds_device_server );
     ~lrs_device_controller();
-    void start_stream( rs2::stream_profile sp, frame_callback_type cb );
-    void stop_stream( rs2_stream stream );
-    void stop_all_streams();
+    void start_streaming( const nlohmann::json & msg );
+    void stop_streaming( const nlohmann::json & msg );
 
 private:
-    class lrs_sensor_streamer;
+    bool find_sensor( const std::string & requested_stream_name, size_t & sensor_index );
 
     rs2::device _rs_dev;
     std::string _device_sn;
-    std::unordered_map<rs2_stream, std::shared_ptr<lrs_sensor_streamer>> stream_to_rs2_sensor;
+    std::vector< rs2::sensor > _sensors;
+    std::shared_ptr< realdds::dds_device_server > _dds_device_server;
 };  // class lrs_device_controller
+
 }  // namespace tools
